Wrap HeroiconsUpdater.RunAsync in try/finally for temp cleanup
Ensures the tmp directory is always cleaned up even when an exception
occurs partway through icon generation or download.

+++ src/HeroiconsGenerator/HeroiconsUpdater.cs
@@ -13,42 +13,50 @@ internal class HeroiconsUpdater(string rootPath, IReleaseProvider releaseProvide
             Directory.Delete(tmpDir, true);
         }
 
-        // Get latest release info
-        Console.WriteLine("Getting latest release info...");
-        var release = await releaseProvider.GetLatestReleaseAsync();
-        Console.WriteLine($"Downloading {release.Version}...");
-
-        var generator = new IconGenerator(rootPath);
-
-        // Update readme badge and version file
-        generator.UpdateReadme(release.Version);
-
-        // Download and extract tarball
-        var optimizedDir = await releaseProvider.DownloadAndExtractAsync(release.TarballUrl, tmpDir);
-
-        Console.WriteLine("Looping through svg files and creating razor components...");
-        var iconSets = new (string iconType, string svgDir)[]
+        try
         {
-            ("Micro", Path.Combine(optimizedDir, "16", "solid")),
-            ("Mini", Path.Combine(optimizedDir, "20", "solid")),
-            ("Solid", Path.Combine(optimizedDir, "24", "solid")),
-            ("Outline", Path.Combine(optimizedDir, "24", "outline")),
-        };
+            // Get latest release info
+            Console.WriteLine("Getting latest release info...");
+            var release = await releaseProvider.GetLatestReleaseAsync();
+            Console.WriteLine($"Downloading {release.Version}...");
 
-        foreach (var (iconType, svgDir) in iconSets)
-            generator.CreateBlazorComponents(iconType, svgDir);
+            var generator = new IconGenerator(rootPath);
 
-        // Generate static registry
-        generator.GenerateHeroiconRegistry(iconSets);
+            // Update readme badge and version file
+            generator.UpdateReadme(release.Version);
 
-        // Update HeroiconName.cs
-        Console.WriteLine("Updating HeroiconName class");
-        generator.UpdateHeroiconNameClass(Path.Combine(optimizedDir, "20", "solid"));
+            // Download and extract tarball
+            var optimizedDir = await releaseProvider.DownloadAndExtractAsync(release.TarballUrl, tmpDir);
 
-        // Clean up
-        Console.WriteLine("Cleaning up temp files...");
-        Directory.Delete(tmpDir, true);
+            Console.WriteLine("Looping through svg files and creating razor components...");
+            var iconSets = new (string iconType, string svgDir)[]
+            {
+                ("Micro", Path.Combine(optimizedDir, "16", "solid")),
+                ("Mini", Path.Combine(optimizedDir, "20", "solid")),
+                ("Solid", Path.Combine(optimizedDir, "24", "solid")),
+                ("Outline", Path.Combine(optimizedDir, "24", "outline")),
+            };
 
-        Console.WriteLine("Done!");
+            foreach (var (iconType, svgDir) in iconSets)
+                generator.CreateBlazorComponents(iconType, svgDir);
+
+            // Generate static registry
+            generator.GenerateHeroiconRegistry(iconSets);
+
+            // Update HeroiconName.cs
+            Console.WriteLine("Updating HeroiconName class");
+            generator.UpdateHeroiconNameClass(Path.Combine(optimizedDir, "20", "solid"));
+
+            Console.WriteLine("Done!");
+        }
+        finally
+        {
+            // Clean up temp directory even if an error occurs
+            if (Directory.Exists(tmpDir))
+            {
+                Console.WriteLine("Cleaning up temp files...");
+                Directory.Delete(tmpDir, true);
+            }
+        }
     }
 }
